% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gate.R
\name{create_gate}
\alias{create_gate}
\title{Create gate object}
\usage{
create_gate(id, label, coords, x_col, y_col, parentgate_col, newgate_col)
}
\arguments{
\item{id}{Could be the newgate_col.}

\item{label}{Could be the newgate_col.}

\item{coords}{List object, which is the return of result2list function.}

\item{x_col}{The same as that in the PolygonGating function.}

\item{y_col}{The same as that in the PolygonGating function.}

\item{parentgate_col}{The same as that in the PolygonGating function.}

\item{newgate_col}{The same as that in the PolygonGating function.}
}
\value{
a S3 gate object.
}
\description{
Create gate object
}
\details{
This function will create a gate object, which is of S3 structure. The gate contains parameters used in interactive polygon drawing and the coordinates of polygon vertexes.
}
