% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GateDecider.R
\name{GateDecider}
\alias{GateDecider}
\title{Create gate column with Polygon Gate}
\usage{
GateDecider(df, gate)
}
\arguments{
\item{df}{the same data dataframe used in PolygonGating function..}

\item{gate}{gate object. returned from PolygonGating function.}
}
\value{
dataframe with a new column for the new gate.
}
\description{
use Polygon Gate to decide if each data point (in parent gate) belongs to the new gate
}
\details{
This function will use the parameter stored in gate to check each points in parent gate if they belong to the new gate. If so, the value in the new column will be True, otherwise be False.
}
\examples{
\dontrun{
# Generate example data
set.seed(123)
n <- 10000
df <- data.frame(
  x1 = runif(n),
  y1 = runif(n),
  value1 = sample(0:99, n, replace = TRUE)
)
df$gate1 <- TRUE

# Perform interactive polygon gating
gate1 <- PolygonGating(
  df = df,
  x_col = "x1",
  y_col = "y1",
  feature_col = "value1",
  parentgate_col = "gate1",
  newgate_col = "gate2"
)

# Apply the gate to the data
df <- GateDecider(gate = gate1, df = df)
}
}
