% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePolygondf.R
\name{GeneratePolygondf}
\alias{GeneratePolygondf}
\title{Generate Polygon dataframe for visualization}
\usage{
GeneratePolygondf(gatelist, gatenames, x_col, y_col)
}
\arguments{
\item{gatelist}{list of gates. list(gate1,gate2, ...)}

\item{gatenames}{names of gates. c("gate1","gate2", ...)}

\item{x_col}{col name of x coordinate. This will be used in ggplot together with data df. So, it should be the same as the col name of x coordinate in the data df.}

\item{y_col}{col name of y coordinate. This will be used in ggplot together with data df. So, it should be the same as the col name of y coordinate in the data df.}
}
\value{
dataframe with gate vertex coordinates and gate names.
}
\description{
use one or multiple gate objects to generate a dataframe, which could be used in ggplot to visualize gates.
}
\details{
This function will transfrom the gates objects into a dataframe, which could be used by ggplot for gates visualization.
}
\examples{
\dontrun{
# Generate example data

x_col= "x1"
y_col= "y1"
feature_col= "value1"

polygon_df = GeneratePolygondf(gatelist = list(gate1,gate2),
                               gatenames = c("gate1","gate2"),
                               x_col,y_col)

ggplot() +
  geom_point(data = df, aes(x = .data[[x_col]],
                            y = .data[[y_col]],
                            color = .data[[feature_col]]),  size = 1) +
  geom_polygon(data = polygon_df,
               aes(x = .data[[x_col]], y = .data[[y_col]],
                   group = gate, fill = gate,), alpha = 0.6) +
  scale_color_gradient(low = "#d6ecf0", high = "#75878a", na.value = NA) +
  scale_fill_brewer(palette = "Pastel1")+
  labs(title = "Scatter Plot with Multiple Polygon Gates", x = "X", y = "Y") +
  theme_minimal()
}
}
