% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result2list.R
\name{result2list}
\alias{result2list}
\title{Reform return from shinyapp into list}
\usage{
result2list(coords_result)
}
\arguments{
\item{coords_result}{The returned coordinates of polygon vertexes from shiny app.}
}
\value{
list object.
}
\description{
Reform return from shinyapp into list
}
\details{
This function will reform return from shinyapp into a list object.
}
\keyword{internal}
