% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data_to_canvas.R
\name{scale_data_to_canvas}
\alias{scale_data_to_canvas}
\title{Scale data for canvas plot}
\usage{
scale_data_to_canvas(df, x_col, y_col, canvas_width, canvas_height)
}
\arguments{
\item{df}{The same as that in PolygonGating function.}

\item{x_col}{The same as that in PolygonGating function.}

\item{y_col}{The same as that in PolygonGating function.}

\item{canvas_width}{The same as that in PolygonGating function.}

\item{canvas_height}{The same as that in PolygonGating function.}
}
\value{
scaled dataframe with additional x_scaled and y_scaled columns.
}
\description{
Scale data for canvas plot
}
\details{
This function will scale coordinates to fit the canvas plot. if all data points have the same x/y coordinates, it will be scaled to be in the center of canvas.
}
\keyword{internal}
